+++ engine/client.cpp
@@ -624,7 +624,7 @@ float CClientState::GetClientInterpAmount()
 		return 0.03f;
 	}
 
-	return s_cl_interp_ratio->GetFloat() * cl_updateinterval->GetFloat();
+	return s_cl_interp_ratio->GetInt() * cl_updateinterval->GetFloat();
 }
 
 //-----------------------------------------------------------------------------
